Move RegionResult into its own directory
