+++ src/NetworkTestWidget/NetworkTestWidget.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Button, Typography } from '@material-ui/core';
 import useTestRunner from './useTestRunner';
-import RegionResult from './RegionResult';
+import RegionResult from './RegionResult/RegionResult';
 import { Region } from '../types';
 
 interface NetworkTestWidgetProps {
